+++ src/core/Utils.cpp
@@ -5,14 +5,14 @@
 
 #include "Utils.h"
 
-long long Utils::getCurrentTimeMillis() {
+unsigned long long Utils::getCurrentTimeMillis() {
     return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
 }
 
-long Utils::getAgeInSecs(long long timestamp) {
+long Utils::getAgeInSecs(unsigned long long timestamp) {
     return (Utils::getCurrentTimeMillis() - timestamp) / 1000;
 }
 
-long long Utils::getMaxTimestampForAge(long age) {
+unsigned long long Utils::getMaxTimestampForAge(double age) {
     return Utils::getCurrentTimeMillis() - (age * 1000);
 }
